# [Core] Scheduler perf fix