@comaniac 2024-04-22
This is a very good example of [`logging-fstring-interpolation (W1203)`](https://pylint.readthedocs.io/en/latest/user_guide/messages/warning/logging-fstring-interpolation.html).

Could you also try this to see if this performance is preserved? If so, we can enable `logging-fstring-interpolation` and `logging-no-lazy` to CI linting.
```
logger.debug("add_seq_group %s", seq_group.request_id)
```