@rkooo567 2024-04-24
@comaniac https://github.com/vllm-project/vllm/pull/4326

Besides, I tried what you suggested on the scheduler, but somehow it is still slower (faster than fstring). So I guess using logger at all is not desirable in the scheduler