@comaniac 2024-04-25
Thanks for trying that. I guess it means logger overhead cannot be ignored in some very intense places.